Add support for parallel readout configurations to UDP config parsing.
This backports the changes made by @GDYendell at #55a89a2 to allow
the UDP configuration to be set up for distributed (via a switch)
or parallel (direct to multiple NICs) networking configuration on
e.g. a 1M detector. Adds the 'all_fems' or by FEM ID mapping of
the destination node UDP config to the JSON format, e.g.:

```
    "nodes": {
        "fem1": [{
            "name": "p29-excalibur01/p1p1",
            "mac": "00:07:43:41:97:c8",
            "ipaddr": "10.0.2.2",
            "port": 61649
        }],
        "fem2": [{
            "name": "p29-excalibur01/p1p2",
            "mac": "00:07:43:41:97:C0",
            "ipaddr": "10.0.2.1",
            "port": 61650
        }]
    },
```